resource_doc 部分接口挪到 resources 模块下
