+++ src/dashboard/apigateway/apigateway/apis/web/resource/doc/urls.py
@@ -1,0 +1,10 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path(
+        path("", views.DocListCreateApi.as_view(), name="resource_doc.list_create"),
+        path("<int:id>/", views.DocUpdateDestroyApi.as_view(), name="resource_doc.update_destroy"),
+    ),
+]
