+++ src/dashboard/apigateway/apigateway/apis/web/resource/urls.py
@@ -1,0 +1,5 @@
+from django.urls import include, path
+
+urlpatterns = [
+    path("<int:resource_id>/docs/", include("apigateway.apis.web.resource.doc.urls")),
+]
