+++ src/dashboard/apigateway/apigateway/apps/support/management/commands/export_public_resource_docs.py
@@ -29,7 +29,7 @@ from django.core.management.base import BaseCommand, CommandError
 from apigateway.apps.support.constants import DocArchiveTypeEnum
 from apigateway.biz.resource_doc.archive_factory import ArchiveFileFactory
 from apigateway.biz.resource_doc.exceptions import NoResourceDocError
-from apigateway.biz.resource_doc.export_doc.generators import DocArchiveGenerator
+from apigateway.biz.resource_doc.exporter.generators import DocArchiveGenerator
 from apigateway.core.constants import GatewayStatusEnum
 from apigateway.core.models import Gateway, Resource
